ref: Rename `utils/registry` for utils only used in the registry target
