+++ src/utils/registry.ts
